+++ server/Data/Dtos/Stock/Response/StockPositionResponseDto.cs
@@ -7,6 +7,6 @@ namespace SolidTradeServer.Data.Dtos.Stock.Response
         public string Isin { get; set; }
         
         public decimal BuyInPrice { get; set; }
-        public int NumberOfShares { get; set; }
+        public decimal NumberOfShares { get; set; }
     }
 }
